+++ pkg/query/storeset.go
@@ -64,7 +64,7 @@ func (s *grpcStoreSpec) Addr() string {
 // Metadata method for gRPC store API tries to reach host Info method until context timeout. If we are unable to get metadata after
 // that time, we assume that the host is unhealthy and return error.
 func (s *grpcStoreSpec) Metadata(ctx context.Context, client storepb.StoreClient) (labelSets []storepb.LabelSet, mint int64, maxt int64, err error) {
-	resp, err := client.Info(ctx, &storepb.InfoRequest{}, grpc.FailFast(false))
+	resp, err := client.Info(ctx, &storepb.InfoRequest{}, grpc.WaitForReady(true))
 	if err != nil {
 		return nil, 0, 0, errors.Wrapf(err, "fetching store info from %s", s.addr)
 	}
@@ -146,9 +146,9 @@ func NewStoreSet(
 		storeNodeConnections:             storeNodeConnections,
 		gRPCInfoCallTimeout:              10 * time.Second,
 		externalLabelOccurrencesInStores: map[string]int{},
-		stores:                make(map[string]*storeRef),
-		storeStatuses:         make(map[string]*StoreStatus),
-		unhealthyStoreTimeout: unhealthyStoreTimeout,
+		stores:                           make(map[string]*storeRef),
+		storeStatuses:                    make(map[string]*StoreStatus),
+		unhealthyStoreTimeout:            unhealthyStoreTimeout,
 	}
 
 	storeNodeCollector := &storeSetNodeCollector{externalLabelOccurrences: ss.externalLabelOccurrences}
@@ -315,7 +315,7 @@ func (s *StoreSet) getHealthyStores(ctx context.Context) map[string]*storeRef {
 				store = &storeRef{StoreClient: storepb.NewStoreClient(conn), cc: conn, addr: addr, logger: s.logger}
 
 				// Initial info call for all types of stores to check gRPC StoreAPI.
-				resp, err := store.StoreClient.Info(ctx, &storepb.InfoRequest{}, grpc.FailFast(false))
+				resp, err := store.StoreClient.Info(ctx, &storepb.InfoRequest{}, grpc.WaitForReady(true))
 				if err != nil {
 					store.close()
 					s.updateStoreStatus(store, err)